Test.QuickCheck.Classes.Semigroup

Description

This module provides:

Support for testing Semigroup subclass instances with QuickCheck.

Reusable properties in the form of Laws definitions.

In general:

Each Laws definition includes properties that relate to just one class.

Laws for superclasses are not transitively included.

Therefore, when testing laws for a particular class, you should make sure to also test laws for all superclasses.

Commutative

commutativeLaws :: forall a. (Arbitrary a, Show a, Eq a, Commutative a) => Proxy a -> Laws

Laws for instances of Commutative.

Tests the following property:

a <> b == b <> a

Reductive

reductiveLaws :: forall a. (Arbitrary a, Show a, Eq a, Reductive a) => Proxy a -> Laws

Laws for instances of Reductive.

Tests the following properties:

a </> b == stripPrefix b a

a </> b == stripSuffix b a

maybe a (b <>) (a </> b) == a

maybe a (<> b) (a </> b) == a

Note that the following superclass laws are not included:

commutativeLaws

leftReductiveLaws

rightReductiveLaws

leftReductiveLaws :: forall a. (Arbitrary a, Show a, Eq a, LeftReductive a) => Proxy a -> Laws

Laws for instances of LeftReductive.

Tests the following properties:

a `isPrefixOf` (a <> b)

isPrefixOf a b == isJust (stripPrefix a b)

maybe b (a <>) (stripPrefix a b) == b

rightReductiveLaws :: forall a. (Arbitrary a, Show a, Eq a, RightReductive a) => Proxy a -> Laws

Laws for instances of RightReductive.

Tests the following properties:

b `isSuffixOf` (a <> b)

isSuffixOf a b == isJust (stripSuffix a b)

maybe b (<> a) (stripSuffix a b) == b

Cancellative

cancellativeLaws :: forall a. (Arbitrary a, Show a, Eq a, Cancellative a) => Proxy a -> Laws

Laws for instances of Cancellative.

Tests the following properties:

(a <> b) </> a == Just b

(a <> b) </> b == Just a

Note that the following superclass laws are not included:

leftCancellativeLaws

rightCancellativeLaws

reductiveLaws

leftCancellativeLaws :: forall a. (Arbitrary a, Show a, Eq a, LeftCancellative a) => Proxy a -> Laws

Laws for instances of LeftCancellative.

Tests the following property:

stripPrefix a (a <> b) == Just b

Note that the following superclass laws are not included:

leftReductiveLaws

rightCancellativeLaws :: forall a. (Arbitrary a, Show a, Eq a, RightCancellative a) => Proxy a -> Laws

Laws for instances of RightCancellative.

Tests the following property:

stripSuffix b (a <> b) == Just a

Note that the following superclass laws are not included:

rightReductiveLaws

GCD

gcdMonoidLaws :: forall a. (Arbitrary a, Show a, Eq a, GCDMonoid a) => Proxy a -> Laws

Laws for instances of GCDMonoid.

Tests the following properties:

gcd a b == commonPrefix a b

gcd a b == commonSuffix a b

isJust (a </> gcd a b)

isJust (b </> gcd a b)

Note that the following superclass laws are not included:

commutativeLaws

reductiveLaws

leftGCDMonoidLaws

rightGCDMonoidLaws

overlappingGCDMonoidLaws

leftGCDMonoidLaws :: forall a. (Arbitrary a, Show a, Eq a, LeftGCDMonoid a) => Proxy a -> Laws

Laws for instances of LeftGCDMonoid.

Tests the following properties:

stripCommonPrefix a b & \(p, _, _) -> p == commonPrefix a b

stripCommonPrefix a b & \(p, x, _) -> p <> x == a

stripCommonPrefix a b & \(p, _, x) -> p <> x == b

stripCommonPrefix a b & \(p, x, _) -> Just x == stripPrefix p a

stripCommonPrefix a b & \(p, _, x) -> Just x == stripPrefix p b

Note that the following superclass laws are not included:

leftReductiveLaws

rightGCDMonoidLaws :: forall a. (Arbitrary a, Show a, Eq a, RightGCDMonoid a) => Proxy a -> Laws

Laws for instances of RightGCDMonoid.

Tests the following properties:

stripCommonSuffix a b & \(_, _, s) -> s == commonSuffix a b

stripCommonSuffix a b & \(x, _, s) -> x <> s == a

stripCommonSuffix a b & \(_, x, s) -> x <> s == b

stripCommonSuffix a b & \(x, _, s) -> Just x == stripSuffix s a

stripCommonSuffix a b & \(_, x, s) -> Just x == stripSuffix s b

Note that the following superclass laws are not included:

rightReductiveLaws

overlappingGCDMonoidLaws :: forall a. (Arbitrary a, Show a, Eq a, OverlappingGCDMonoid a) => Proxy a -> Laws

Laws for instances of OverlappingGCDMonoid.

Tests the following properties:

overlap a b <> stripPrefixOverlap a b == b

stripSuffixOverlap b a <> overlap a b == a

stripOverlap a b & \(_, x, _) -> x == overlap a b

stripOverlap a b & \(_, _, x) -> x == stripPrefixOverlap a b

stripOverlap a b & \(x, _, _) -> x == stripSuffixOverlap b a

Note that the following superclass laws are not included:

leftReductiveLaws

rightReductiveLaws

cancellativeGCDMonoidLaws :: forall a. (Arbitrary a, Show a, Eq a, Cancellative a, GCDMonoid a) => Proxy a -> Laws

Laws for instances of Cancellative and GCDMonoid.

Tests the following properties:

gcd (a <> b) (a <> c) == a <> gcd b c

gcd (a <> c) (b <> c) == gcd a b <> c

Note that the following superclass laws are not included:

cancellativeLaws

gcdMonoidLaws

Monus

monusLaws :: forall a. (Arbitrary a, Show a, Eq a, Monus a) => Proxy a -> Laws

Laws for instances of Monus.

Tests the following properties:

a <\> b == stripPrefixOverlap b a

a <\> b == stripSuffixOverlap b a

Note that the following superclass laws are not included:

commutativeLaws

overlappingGCDMonoidLaws

Null

monoidNullLaws :: forall a. (Arbitrary a, Show a, Eq a, MonoidNull a) => Proxy a -> Laws

Laws for instances of MonoidNull.

Tests the following property:

null a == (a == mempty)

Positive

positiveMonoidLaws :: forall a. (Arbitrary a, Show a, Eq a, PositiveMonoid a) => Proxy a -> Laws

Laws for instances of PositiveMonoid.

Tests the following properties:

null (a <> b) == (null a && null b)

Note that the following superclass laws are not included:

monoidNullLaws

Group

groupLaws :: forall a. (Arbitrary a, Show a, Eq a, Group a) => Proxy a -> Laws

Laws for instances of Group.

Tests the following properties:

invert mempty == mempty

a <> invert a == mempty

invert a <> a == mempty

a ~~ mempty == a

a ~~ a == mempty

a ~~ b == a <> invert b

n >= 0 ==> pow a n == mconcat (replicate n a)

n <= 0 ==> pow a n == invert (mconcat (replicate (abs n) a))